+++ .travis.yml
@@ -6,28 +6,16 @@ jdk:
 install:
   - echo y | android update sdk -u -a -t tools
   - echo y | android update sdk -u -a -t platform-tools
-  - echo y | android update sdk -u -a -t build-tools-25.0.3
-  - echo y | android update sdk -u -a -t android-25
+  - echo y | android update sdk -u -a -t build-tools-28.0.3
+  - echo y | android update sdk -u -a -t android-28
   - echo y | android update sdk -u -a -t extra-google-m2repository
-  - echo y | android update sdk -u -a -t extra-android-m2repository
+  - echo y | android update sdk -u -a -t extra-android-support
+  - echo y | android update sdk -u -a -t extra-google-google_play_services
 
 script:
   - ./gradlew lint
   - ./gradlew checkstyle
   - ./gradlew pmd
-  - ./gradlew buld
+  - ./gradlew test -i
 
-branches:
-  except:
-    - gh-pages
 
-sudo: false
-
-before_cache:
-  - rm -f  $HOME/.gradle/caches/modules-2/modules-2.lock
-  - rm -fr $HOME/.gradle/caches/*/plugin-resolution/
-cache:
-  directories:
-    - $HOME/.gradle/caches/
-    - $HOME/.gradle/wrapper/
-    - $HOME/.android/build-cache
