+++ app/build.gradle
@@ -24,7 +24,10 @@ android {
             xmlOutput file("$project.buildDir/reports/lint/lint-result.xml")
         }
 
+
+
     }
+
 
     buildTypes {
         release {
@@ -48,6 +51,7 @@ dependencies {
 
     //dagger 2
     implementation 'com.google.dagger:dagger:2.28.1'
+    testImplementation 'junit:junit:4.13'
     annotationProcessor 'com.google.dagger:dagger-compiler:2.28.1'
     implementation 'androidx.annotation:annotation:1.1.0'
