+++ lib/build.gradle
@@ -25,6 +25,13 @@ android {
 
     }
 
+    sourceSets {
+        test {
+            resources { srcDir 'test/birthdaytest' }
+        }
+    }
+
+
     buildTypes {
         release {
             minifyEnabled false
@@ -35,6 +42,7 @@ android {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
+
 }
 
 dependencies {
